+++ pl_bolts/models/self_supervised/cpc/cpc_module.py
@@ -11,7 +11,6 @@ import torch
 import torch.optim as optim
 from pytorch_lightning.utilities import rank_zero_warn
 
-from pl_bolts.callbacks.self_supervised import SSLOnlineEvaluator
 from pl_bolts.losses.self_supervised_learning import CPCTask
 from pl_bolts.models.self_supervised.cpc.networks import cpc_resnet101
 from pl_bolts.models.self_supervised.cpc.transforms import (
@@ -230,6 +229,7 @@ class CPCV2(pl.LightningModule):
 
 
 def cli_main():
+    from pl_bolts.callbacks.ssl_online import SSLOnlineEvaluator
     from pl_bolts.datamodules import CIFAR10DataModule, STL10DataModule
     from pl_bolts.datamodules.ssl_imagenet_datamodule import SSLImagenetDataModule
